Improve the performance of the delivery system
By moving the http requests into a goroutine, performance has increased
dramatically.  Processing 1000 postbacks used to take upwards of 12
minutes, but now it can take as little as 2 seconds.  Additionally, a
high performance logger (zap) is in use in order to speed up
performance, if only by a marginal amount.