Split the ingestion service docker image
Instead of using a premade php image that ran both nginx AND php-fpm,
use a seperate nginx image to reverse proxy and load balance the php-fpm
service.  The benefit is two-fold.  The containers take far less time to
start because the composer install process happens when the image is
built, not when the container is started.  Additionally, the php portion
can now be properly load balanced without spinning up an nginx instance
for every single php instance.

However, this does come with one (minor) drawback: the virtual host at
which the ingestion service will be accessed must be specified in the
environment file.  In this commit, it is set to `localhost`.

Should fix #1